+++ R/sRCT.R
@@ -1,32 +1,25 @@
-#sRCT
-#Todowith real data <-
-
 # ==== DOCUMENTATION ====
 
 #' simulated Randomised Clinical Trial (sRCT)
 #'
-#' `sRCT()` is a function which simulates a randomised clinical trial with a binary outome and returns a dataframe. This version is validated to be used for analysis of interaction in a factorial design.
+#' `sRCT()` is a function which simulates a randomised clinical trial with a binary outcome and returns a dataframe. This version is validated to be used for analysis of interaction in a factorial design.
 #'
 #' @name sRCT
 #'
-#' @usage sRCT(part_tbl = NULL, all_sizes = NULL,
-#' n_pop = 100,n_sites = 1,design = c(2,2,2),
-#' rrr = c(0.05,0.05,0), interaction = c(`1<-2` = 0.05, `1<>2` = -0.05),
-#' strata_var = c("age","sex"), strata_site = T,
-#' strata_risk = c(age=0.3,sex=0.5),
-#' outcome_risk = 0.492)
+#' @usage sRCT(all_sizes = NULL, n_pop = 1000,
+#' n_sites = 10, design = c(2,2,2),
+#' rrr = c(0.05,0.05,0), outcome_risk = 0.492,
+#' interaction = c(`1-2` = 0.05, `1-2-3` = -0.05),
+#' site_re = 0.05)
 #'
-#' @param part_tbl Here a participation data frame should be imported. `[TODO: NOT FUNCTIONAL]`
 #' @param all_sizes Size of blocks in allocation table. If left empty the three lowest possible block sizes will be randomly assigned.
 #' @param n_pop Number of participants included in the trial.
 #' @param n_sites Number of sites
-#' @param design Number of sites as a list where each element corresponds to an intervention and the number in the element is the number of groups. So for a 2x2 factorial design `c(2,2)` should be used. `[TODO: THREE GROUPS]`
-#' @param rrr relative risk reduction for each intervention so for the abovementioned 2x2 factorial desing with RRR of 0.05 and 0.10 we would use `c(0.05,0.10)`.
-#' @param interaction Interaction between interventions with a named list. If intervention 2 increases the RRR of intervention 1 we would use `1<-2 = 0.05`.
-#' @param strata_var Variable which would be used for stratification.
-#' @param strata_site If randomisation should be stratified by site
-#' @param strata_risk The frequency of a dichotomised strata. Named list where the name must correspond to a strata var.
-#' @param outcome_risk The baseline risk of the dichotomous primary outcome.
+#' @param design Number of sites as a list where each element corresponds to an intervention and the number in the element is the number of groups. So for a 2x2 factorial design `c(2,2)` should be used.
+#' @param rrr relative risk reduction for each intervention so for the abovementioned 2x2 factorial design with RRR of 0.05 and 0.10 we would use `c(0.05,0.10)`.
+#' @param outcome_risk The baseline risk (probability in absolute percentage) of the dichotomous primary outcome.
+#' @param interaction Interaction between interventions with a named list. If interaction exists between intervention 1 and 2 we would use `1-2 = 0.05`.
+#' @param site_re The size of the random effect of site, default is `0.05`.
 #'
 #' @return Returns a dataframe with an individual participant data frame.
 #'
@@ -35,157 +28,107 @@
 #' @examples sRCT()
 #'
 #' @importFrom stats rbinom
+#' @importFrom stats rnorm
 #' @export
 #
 # ==== FUNCTION ====
 
 # #ADD PERCENTAGE OF BLOCKS
-# part_tbl = NULL; all_sizes = NULL;
+# all_sizes = NULL;
 # n_pop = 2000; #Antal inkluderede
 # n_sites = 10; #Antal sites
 # design = c(2,2,2); #No. Interventioner / grupper
 # rrr = c(0.05,0.05,-.05); #RRR
-# interaction = c(`1<-2` = 0.05);
-# strata_var = c("age","sex"); #Strata variables
-# strata_site = T; #Stratify by site
-# strata_risk = c(age=0.3,sex=0.5); #Name and Risk of strata
+# interaction = c(`1-2` = 0.05, `1-2-3` = -0.05);
 # outcome_risk = 0.49
 
-sRCT <- function(part_tbl = NULL, all_sizes = NULL,
-                 n_pop = 100, #Antal inkluderede
-                 n_sites = 1, #Antal sites
-                 design = c(2,2,2), #No. Interventioner / grupper
-                 rrr = c(0.05,0.05,0), #RRR
-                 interaction = c(`1<-2` = 0.05, `1<>2` = -0.05),
-                 strata_var = c("age","sex"), #Strata variables
-                 strata_site = T, #Stratify by site
-                 strata_risk = c(age=0.3,sex=0.5), #Name and Risk of strata
-                 outcome_risk = 0.492){
-   #Generate sites ----
-   if(length(n_sites) == 1){
+sRCT <- function(all_sizes = NULL, n_pop = 1000,
+                 n_sites = 10, design = c(2,2,2),
+                 rrr = c(0.05,0.05,0), outcome_risk = 0.492,
+                 interaction = c(`1-2` = 0.05, `1-2-3` = -0.05),
+                 site_re = 0.05){
+
+   trial <- NULL
+
+   #Generate site names ----
       site_id <- replicate(n_sites, paste0(paste(sample(LETTERS,5, replace=TRUE),collapse=""),paste(sample(c(0:9),2, replace=TRUE),collapse="")))
-   }else{
-      site_id <- unique(part_tbl$site)
-   }
 
-   #Generate block ----
-   genBlock_tbl <- function(design){
-      expand_grid_list <- NULL
-      for(i in 1:length(design)){
-         expand_grid_list <- c(expand_grid_list, list(0:(design[i]-1)))
+   #Generate site probability ----
+      site_prop <- rnorm(n_sites,mean=10,sd=7)
+      site_prop[site_prop < 0] <- 0
+      while(any(round(site_prop,digits=2) == 0)){
+         site_prop[round(site_prop,digits=2) == 0 ] <- 0.02
+         site_prop <- site_prop/sum(site_prop)
       }
-      block_tbl <- expand.grid(expand_grid_list)
-      return(block_tbl)
-   }
-   block_tbl <- genBlock_tbl(design)
 
-   #Generate participants table ----
-   genPart_tbl <- function(n_pop, site_id, strata_var, strata_risk, outcome_risk){
-      part_tbl <- NULL
-      if(!is.null(site_id)){
-         part_tbl$site <- sample(site_id,n_pop,replace=TRUE)
-      }
-      if(!is.null(strata_var)){
-         for(i in 1:length(strata_var)){
-            part_tbl[[names(strata_risk[i])]] <- sample(0:1,n_pop,replace=T,prob=c(1-strata_risk[i],strata_risk[i]))
+   #Generate blocks ----
+      genBlock_tbl <- function(design){
+         expand_grid_list <- NULL
+         for(i in 1:length(design)){
+            expand_grid_list <- c(expand_grid_list, list(0:(design[i]-1)))
          }
+         block_tbl <- expand.grid(expand_grid_list)
+         return(block_tbl)
       }
-      part_tbl <- data.frame(part_tbl)
-      part_tbl$outcome <- sample(0:1,nrow(part_tbl),replace=T,prob=c(outcome_risk,1-outcome_risk))
-      return(part_tbl)
-   }
-   if(is.null(part_tbl)) part_tbl <- genPart_tbl(n_pop,site_id,strata_var,
-                                                 strata_risk, outcome_risk)
+      block_tbl <- genBlock_tbl(design)
 
-   #Generate allocation table ----
-   genAll_tbl <- function(n_pop, site_id, block_tbl,strata_site, strata_var,
-                          all_sizes = NULL){
+   #Generate participants ----
+      #Allocate participants to trials
+      trial$site <- sample(site_id,n_pop,replace=T,prob=site_prop)
+      trial <- data.frame(trial)
+      for(i in colnames(block_tbl)) trial[[i]] <- NA
+
+      #Find or define allocation sizes
       if(is.null(all_sizes)) {
          all_sizes <- c(nrow(block_tbl),nrow(block_tbl)*2,nrow(block_tbl)*3)
       }
 
-      #Function to generate allocation for one strata
-      genAll_simple <- function(n_pop, block_tbl, all_sizes){
-         results <- NULL
-         k <- 0
+      #Allocate participants to interventions
+      for(i in unique(trial$site)){
+         tmp_n <- sum((trial$site == i)*1)
          all_size <- NULL
-         while(sum(all_size) <= n_pop){
+         while(sum(all_size) <= tmp_n){
             all_size <- c(all_size,sample(all_sizes,1))
          }
+         alloc_seq <- NULL
          for(j in 1:length(all_size)){
             block <- do.call("rbind",
                              replicate(all_size[j]/nrow(block_tbl),
                                        block_tbl, simplify = FALSE))
-            block <- data.frame(block_tbl[sample(c(1:nrow(block_tbl)),
-                                                 nrow(block_tbl)),])
-            results <- rbind(results,block)
+            block <- data.frame(block[sample(c(1:nrow(block)),
+                                                 nrow(block)),])
+            alloc_seq <- rbind(alloc_seq,block)
          }
-         k <- nrow(results)
-         colnames(results) <- paste0("Var", 1:ncol(results))
-         return(results)
+         trial[trial$site == i,colnames(alloc_seq)] <- alloc_seq[c(1:tmp_n),]
+
+         #Define site specific outcome risk
+         trial$rr_site[trial$site == i] <- rnorm(1,mean=log(outcome_risk),sd=site_re)
       }
 
-      strata <- NULL
-      for(i in strata_var){
-         strata[[i]] <- c(0,1)
+      #Find RRR for each intervention
+      for(i in c(1:(length(rrr)))){
+         rrr_temp <- trial[[paste0("Var",i)]]*rrr[i]
+         trial[[paste0("rr_intervention",i)]] <- log(1-replace(rrr_temp,is.na(rrr_temp*rrr[i]),0))
       }
-      if(strata_site){
-         strata[["site"]] <- site_id
-      }
-      strata <- data.frame(expand.grid(strata))
-      all_tbl <- NULL
-      for(i in 1:nrow(strata)){
-         temp <- suppressWarnings(cbind(strata[i,],genAll_simple(n_pop,block_tbl,all_sizes)))
-         colnames(temp)[1:ncol(strata)] <- colnames(strata)
-         all_tbl <- rbind(all_tbl,temp)
-      }
-      return(all_tbl)
-   }
-   all_tbl <- genAll_tbl(n_pop, site_id, block_tbl,strata_site, strata_var,
-                         all_sizes)
 
-   #Calculate RRR ----
-   for(i in 1:length(rrr)){
-      col <- colnames(all_tbl)[grepl("Var",colnames(all_tbl))][i]
-      all_tbl[[paste0("rrr",i)]] <- 0
-      all_tbl[[paste0("rrr",i)]][all_tbl[[col]] == 1] <- rrr[i]
-   }
-   #Add interactions
-   if(!is.null(interaction)){
-      for(i in 1:length(interaction)){
-         int_set <- names(interaction)[i]
-         int_eff <- interaction[i]
-         if(grepl("<-",int_set)){
-            all_tbl[[paste0("rrr",substr(int_set,1,1))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1] <- all_tbl[[paste0("rrr",substr(int_set,1,1))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1]+all_tbl[[paste0("rrr",substr(int_set,1,1))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1]*int_eff
-         }
-         if(grepl("<>",int_set)){
-            all_tbl[[paste0("rrr",substr(int_set,1,1))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1 & all_tbl[[paste0("Var",substr(int_set,1,1))]] == 1] <- all_tbl[[paste0("rrr",substr(int_set,1,1))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1 & all_tbl[[paste0("Var",substr(int_set,1,1))]] == 1]+all_tbl[[paste0("rrr",substr(int_set,1,1))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1 & all_tbl[[paste0("Var",substr(int_set,1,1))]] == 1]*int_eff
-            all_tbl[[paste0("rrr",substr(int_set,4,4))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1 & all_tbl[[paste0("Var",substr(int_set,1,1))]] == 1] <- all_tbl[[paste0("rrr",substr(int_set,4,4))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1 & all_tbl[[paste0("Var",substr(int_set,1,1))]] == 1]+all_tbl[[paste0("rrr",substr(int_set,4,4))]][all_tbl[[paste0("Var",substr(int_set,4,4))]] == 1 & all_tbl[[paste0("Var",substr(int_set,1,1))]] == 1]*int_eff
+      #Find RRR for each interaction TO ADD
+      if(length(interaction) > 0){
+         for(i in c(1:(length(interaction)))){
+            tmp_setting <- names(interaction)[i]
+            formel <- NULL
+            for(j in unlist(unique(strsplit(tmp_setting,"-")))){
+               formel <- paste0(formel,"trial[[\"Var",j,"\"]]*")
+            }
+            formel <- paste0(formel,"log(1-",interaction[i],")")
+            trial[[paste0("rr_interaction",i)]] <- eval(parse(text = formel))
+
          }
       }
-   }
-   all_tbl$rrr <- rowSums(all_tbl[,grepl("rrr",colnames(all_tbl))])
 
-
-   #Include participants ----
-   '%!in%' <- function(x,y)!('%in%'(x,y))
-   trial <- part_tbl[sample(1:nrow(part_tbl),n_pop,replace=T),]
-   colz <- colnames(trial)[colnames(trial) != "outcome"]
-   temp <- NULL
-   for(i in unique(interaction(trial[,colz]))){
-      n_all <- sum((i == interaction(trial[,colz]))*1)
-      test <- cbind(trial[interaction(trial[,colz]) == i,],all_tbl[interaction(all_tbl[,colz]) %in% i,][1:n_all,which(colnames(all_tbl) %!in% colz)])
-      temp <- rbind(temp,test)
-   }
-   trial <- temp
-
-   re_rrr <- trial$rrr[trial$rrr > 0 & trial$outcome == 1]
-   re_rrr <- 1-re_rrr
-   trial$outcome[trial$rrr > 0 & trial$outcome == 1] <- rbinom(length(re_rrr), size = 1, prob=re_rrr)
-
-   re_rrr <- trial$rrr[trial$rrr < 0 & trial$outcome == 0]
-   re_rrr <- abs(re_rrr)
-   trial$outcome[trial$rrr < 0 & trial$outcome == 0] <- rbinom(length(re_rrr), size = 1, prob=re_rrr)
+   #Generate outcome ----
+      trial$outcome_prob <- exp(rowSums(trial[grepl("rr_",colnames(trial))]))
+      trial$outcome <- lapply(trial$outcome_prob, function(x) sample(c(0,1),1,prob = c(1-x,x)))
+      trial$outcome <- as.numeric(trial$outcome)
 
    return(trial)
 }